+++ config/users.php
@@ -1,0 +1,9 @@
+<?php
+/**
+ * Application-specific configuration for CakeDC Users plugin
+ *
+ * @see https://github.com/CakeDC/users/blob/15.next-cake5/config/users.php
+ */
+
+return [
+];