+++ CHANGELOG.md
@@ -11,6 +11,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - Set up application based on CakePHP 5
 - Add this Changelog #6
+- Implement authentication with registration, login and logout through cakedc/users plugin #4 #5
 
 ### Changed